ai_crew_tutor/requirements.txt
typing_extensions>=4.9.0
streamlit
crewai
pyyaml
sentence-transformers
pysqlite3-binary
langchain
langchain-community
langchain-openai
langchain-huggingface
chromadb>=0.4.22
faiss-cpu
numpy
scikit-learn
openai
python-dotenv
groq
langchain-groq
cohere
requests
huggingface
plotly
streamlit-analytics
streamlit-modal
firebase-admin
